requirements.txt
tensorflow>=2.16.0
numpy>=1.26.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.12.2
tensorboard>=2.16.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

